W4D1/Q2/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
scikit-learn==1.3.2
numpy==1.26.0
pytest==7.4.3 